strip go binaries to reduce OCI size
The go binaries produced by the build can be stripped which should
reduce the overall size of the OCI produced:

$ docker images
REPOSITORY   TAG       IMAGE ID       CREATED              SIZE
vault        1.15.0    6f62316c5777   About a minute ago   284MB

This method is used in the vault snap to achieve the same optimisation.

Set source-depth to 1 to reduce the time taken to clone the upstream
repo (it has alot of history).